refactor(memory-nodes): invoke handlers directly, remove IPC/route path
Per dvargasfuertes review: routing through the daemon main thread via IPC
causes instability, and these handlers can be called in-process.

- nodes.ts: transport ipc→local, drop cliIpcCall/exitFromIpcResult,
  call handleListMemory/handleDeleteMemory/handleUpdateMemory directly
- Delete memory-cli-routes.ts (no longer needed)
- Remove MEMORY_CLI_ROUTES from runtime/routes/index.ts
- Remove /v1/memory/{delete,list,update} from openapi.yaml
- Extend cli-no-daemon-internals local allowlist for graph/tool-handlers
