+++ assistant/eslint-rules/cli-no-daemon-internals.js
@@ -67,6 +67,10 @@ const ALLOWED_PREFIXES = {
     // commands/memory/ nesting.
     "../../plugins/defaults/memory/memory-retrospective-job",
     "../../../plugins/defaults/memory/memory-retrospective-job",
+    // Memory nodes CRUD — `memory nodes list/delete/update` invoke the graph
+    // tool-handlers directly in-process (no IPC, no daemon dependency).
+    // Depth-3 for commands/memory/ nesting.
+    "../../../plugins/defaults/memory/graph/tool-handlers",
     // Standalone tool execution — `tools run` executes a single tool
     // in-process from the filesystem (no daemon, no IPC), so it imports the
     // standalone runner directly. Depth-2 for commands/ nesting.
